Move drb fetch to server directory
