+++ pages/api/drb.ts
@@ -1,33 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next"
 
-import type { SearchParams } from "../../src/types/searchTypes"
-import type { DRBResults } from "../../src/types/drbTypes"
-import nyplApiClient from "../../src/server/nyplApiClient/index"
-import { DRB_API_NAME } from "../../src/config/constants"
-import { getDRBQueryStringFromSearchParams } from "../../src/utils/drbUtils"
 import { mapQueryToSearchParams } from "../../src/utils/searchUtils"
-
-/**
- * Getter function for DRB results. Accepts a SearchParams object as an argument.
- */
-export async function fetchDRBResults(
-  searchParams: SearchParams
-): Promise<DRBResults | Error> {
-  const drbQueryString = getDRBQueryStringFromSearchParams(searchParams)
-
-  try {
-    const client = await nyplApiClient({ apiName: DRB_API_NAME })
-    const { data } = await client.get(drbQueryString)
-
-    return {
-      works: data.works,
-      totalWorks: data.totalWorks,
-    }
-  } catch (error) {
-    console.log(`Error fetching DRB results ${error.message}`)
-    throw new Error(error)
-  }
-}
+import { fetchDRBResults } from "../../src/server/api/drb"
 
 /**
  * Default API route handler for ResearchNow DRB
